Revisa todos os componetes para entrega da sprint 1

+++ frontend/src/components/Tabelas/CelulaTabelaLivros/index.jsx
@@ -1,34 +1,33 @@
 import React, { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import styles from "./index.module.scss";
-import { IoPencil, IoTrash, IoTrashOutline } from "react-icons/io5";
-import ModalExcluir from "../../Modal/ModalExcluir";
+import { IoPencil, IoTrash } from "react-icons/io5";
 import ModalEditar from "../../Modal/ModalEditar";
-import { api } from '../../../config/api';
+import Input from "../../Inputs/Input";
+import { api } from "../../../config/api";
 import { useAuth } from "../../../context/AuthContext";
+import { useForm } from "react-hook-form";
 
-const CelulaTabelaLivros = ({ livro, onDelete  }) => {
+const CelulaTabelaLivros = ({ livro, setModalMensagemAberto, }) => {
   const [livroSelecionado, setLivroSelecionado] = useState(null);
   const [modalEditarAberto, setModalEditarAberto] = useState(false);
-  const [modalAberto, setModalAberto] = useState(false);
-  const {token} = useAuth()
-  
+  const [isDeleteModalOpen, setIsDeleteModalOpen] = useState(false);
+  const [isPasswordModalOpen, setIsPasswordModalOpen] = useState(false);
+  const [password, setPassword] = useState("");
+  const [passwordMessage, setPasswordMessage] = useState(null);
+
+  const { token } = useAuth();
   const navigate = useNavigate();
 
-
-  const abreModal = (livroRelacionado) => {
-    setLivroSelecionado(livroRelacionado);
-    setModalAberto(true);
-  };
+  const {
+    register,
+    handleSubmit,
+    formState: { errors },
+  } = useForm();
 
   const showModal = (livroRelacionado) => {
     setLivroSelecionado(livroRelacionado);
     setModalEditarAberto(true);
-  };
-
-  const fechaModal = () => {
-    setModalAberto(false);
-    setLivroSelecionado(null);
   };
 
   const closeModal = () => {
@@ -36,19 +35,50 @@ const CelulaTabelaLivros = ({ livro, onDelete  }) => {
     setLivroSelecionado(null);
   };
 
-  const deletarLivro = async () => {
+  const handleDeleteClick = () => {
+    setLivroSelecionado(livro);
+    setIsDeleteModalOpen(true);
+  };
+
+  const openPasswordModal = () => {
+    setIsDeleteModalOpen(false);
+    setIsPasswordModalOpen(true);
+    setPasswordMessage(null);
+  };
+
+  const closeDeleteModal = () => {
+    setIsDeleteModalOpen(false);
+  };
+
+  const closePasswordModal = () => {
+    setIsPasswordModalOpen(false);
+    setPassword("");
+  };
+
+  const handleConfirmDelete = async (data) => {
     try {
-      const response = await api.delete(`/livros/${livro.id}`, {
+      const response = await api.get(`/check-senha?password=${data.password}`, {
         headers: {
-          Authorization: `Bearer ${token}`
-        }
+          Authorization: `Bearer ${token}`,
+        },
       });
-      console.log("Livro excluído com sucesso:", response.data);
-      onDelete(livro.id);
-      fechaModal();
-      buscaLivro(); // Atualiza a lista após exclusão
+
+      if (response.data.status) {
+        await api.delete(`/livros/${livro.id}`, {
+          headers: {
+            Authorization: `Bearer ${token}`,
+          },
+        });
+
+        onDelete(livro.id);
+        closePasswordModal();
+        setMessage("Livro excluído com sucesso");
+      setModalMensagemAberto(true);
+      } else {
+        setPasswordMessage("Senha incorreta");
+      }
     } catch (error) {
-      console.error("Erro ao excluir o livro:", error);
+      console.error("Erro ao validar senha:", error);
     }
   };
 
@@ -82,15 +112,86 @@ const CelulaTabelaLivros = ({ livro, onDelete  }) => {
         <div className={styles.num}>{livro.numRegistro}</div>
         <div className={styles.opcoes}>
           <div className={styles.editar} onClick={() => showModal(livro)}>
-            <IoPencil/>
+            <IoPencil />
           </div>
-          <div className={styles.excluir} onClick={() => abreModal(livro)}>
-            <IoTrash/>
+          <div className={styles.excluir} onClick={handleDeleteClick}>
+            <IoTrash />
           </div>
         </div>
       </div>
-      <ModalExcluir textoModal={`Tem certeza de que deseja excluir o livro ${livro.nome}`} onClick={deletarLivro} modalAberto={modalAberto} fechaModal={fechaModal} itemSelecionado={livroSelecionado}/>
-      <ModalEditar closeModal={closeModal} modalEditarAberto={modalEditarAberto} showModal={showModal} livro={livro}/>
+
+      {/* Modal de Edição */}
+      <ModalEditar
+        closeModal={closeModal}
+        modalEditarAberto={modalEditarAberto}
+        showModal={showModal}
+        livro={livro}
+      />
+
+      {/* Modal de Confirmação de Exclusão */}
+      {isDeleteModalOpen && (
+        <div className={styles.modal}>
+          <div className={styles.modalExcluir}>
+            <h3 className={styles.titulo}>Excluir Livro</h3>
+            <p className={styles.mensagem}>
+              Tem certeza de que deseja excluir o livro{" "}
+              <strong>{livro.nome}</strong>?
+            </p>
+            <div className={styles.botoes}>
+              <button
+                onClick={openPasswordModal}
+                className={styles.deleteButton}
+              >
+                Excluir
+              </button>
+              <button
+                onClick={closeDeleteModal}
+                className={styles.closeButton}
+              >
+                Cancelar
+              </button>
+            </div>
+          </div>
+        </div>
+      )}
+
+      {/* Modal de Confirmação de Senha */}
+      {isPasswordModalOpen && (
+        <div className={styles.modal}>
+          <form
+            onSubmit={handleSubmit(handleConfirmDelete)}
+            className={styles.modalSenha}
+          >
+            <h3 className={styles.titulo}>Confirmar Exclusão</h3>
+            <p className={styles.mensagem}>Por favor, digite sua senha:</p>
+              <p className={styles.error}>{passwordMessage && (passwordMessage)}</p>
+            <div>
+              <label htmlFor="senha">Senha</label>
+              <Input
+                type="password"
+                placeholder="Digite sua senha"
+                value={password}
+                onChange={(value) => setPassword(value)}
+                {...register("password", {
+                  required: "A senha é obrigatória",
+                })}
+              />
+                <p className={styles.error}>{errors.password && (errors.password.message)}</p>
+            </div>
+            <div className={styles.botoes}>
+              <button type="submit" className={styles.saveButton}>
+                Confirmar
+              </button>
+              <button
+                onClick={closePasswordModal}
+                className={styles.closeButton}
+              >
+                Cancelar
+              </button>
+            </div>
+          </form>
+        </div>
+      )}
     </div>
   );
 };
